requirements.txt
pandas
pyarrow
numpy
scipy
typing
transformers
huggingface_hub
pytorch
torchtext
pytorch-lightning
torchmetrics
plotly